New look and new content
Change the theme to Arabica which is white background and much easier on the eyes.

Add two new posts:
  * A brief musical journey
  * Making again

Changed the repository to use git-lfs to store the image assets.

TODO: extend the Makefile with exiftool stripping of metadata
TODO: consider using it as git pre-commit hook to ensure compliance