Update build-master.yml
Change aws ACL to public-read.

CF distributions using private access identities are limited to only using a single index.html at the root of the bucket and not in any subdirectory as is usually the case in public S3 websites. We'll revert to public-read S3 website with CF caching the S3 website behavior.